barati/requirements.txt
click==6.2
Django==1.8.4
django-axes==1.4.0
django-braces==1.8.1
django-haystack==2.4.1
django-model-utils==2.4
django-star-ratings==0.4.0
djangorestframework==3.1.1
elasticsearch==1.9.0
oauthlib==1.0.3
psycopg2==2.6
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.13
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
urllib3==1.13
wheel==0.24.0
